+++ fe2o3-amqp-types/src/messaging/format/map_builder.rs
@@ -1,44 +1,45 @@
 //! AnnotationBuilder for types that are simply a wrapper around Annotation
 
-use std::{collections::BTreeMap, marker::PhantomData};
+use std::{hash::Hash, marker::PhantomData};
 
-use serde_amqp::{primitives::Symbol, Value};
+use serde_amqp::{primitives::OrderedMap, Value};
 
 use crate::primitives::SimpleValue;
 
-use super::{ApplicationProperties, DeliveryAnnotations, Footer, MessageAnnotations};
+use super::{
+    annotations::OwnedKey, ApplicationProperties, DeliveryAnnotations, Footer, MessageAnnotations,
+};
 
 /// Builder for types that are simply a wrapper around map
 /// ([`DeliveryAnnotations`], [`MessageAnnotations`], [`Footer`], [`ApplicationProperties`])
 ///
 /// This simply provides a convenient way of inserting entries into the map
 #[derive(Debug)]
-pub struct MapBuilder<K, V, T>
-where
-    K: Ord,
-{
-    map: BTreeMap<K, V>,
+pub struct MapBuilder<K, V, T> {
+    map: OrderedMap<K, V>,
     marker: PhantomData<T>,
 }
 
-impl<K: Ord, V, T> Default for MapBuilder<K, V, T> {
+impl<K, V, T> Default for MapBuilder<K, V, T> {
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl<K, V, T> MapBuilder<K, V, T>
-where
-    K: Ord,
-{
+impl<K, V, T> MapBuilder<K, V, T> {
     /// Creates a new builder for annotation types
     pub fn new() -> Self {
         Self {
-            map: BTreeMap::new(),
+            map: OrderedMap::new(),
             marker: PhantomData,
         }
     }
+}
 
+impl<K, V, T> MapBuilder<K, V, T>
+where
+    K: Hash + Eq,
+{
     /// A convenience method to insert an entry into the annotation map
     pub fn insert(mut self, key: impl Into<K>, value: impl Into<V>) -> Self {
         self.map.insert(key.into(), value.into());
@@ -46,21 +47,21 @@ where
     }
 }
 
-impl MapBuilder<Symbol, Value, DeliveryAnnotations> {
+impl MapBuilder<OwnedKey, Value, DeliveryAnnotations> {
     /// Build [`DeliveryAnnotations`]
     pub fn build(self) -> DeliveryAnnotations {
         DeliveryAnnotations(self.map)
     }
 }
 
-impl MapBuilder<Symbol, Value, MessageAnnotations> {
+impl MapBuilder<OwnedKey, Value, MessageAnnotations> {
     /// Build [`MessageAnnotations`]
     pub fn build(self) -> MessageAnnotations {
         MessageAnnotations(self.map)
     }
 }
 
-impl MapBuilder<Symbol, Value, Footer> {
+impl MapBuilder<OwnedKey, Value, Footer> {
     /// Build [`Footer`]
     pub fn build(self) -> Footer {
         Footer(self.map)