+++ fe2o3-amqp/src/link/sender.rs
@@ -237,9 +237,9 @@ impl Sender {
     ///
     /// - Any type that implements [`serde::Serialize`] will be implicitly wrapped inside an [`AmqpValue`].
     /// (**BREAKING** change in v0.5.0: `AmqpValue`, `AmqpSequence` and `Data` no longer implement
-    /// `Serialize` unless placed in a `Serializable<T>` wrapper, and thus they can be used to specify body section type. 
+    /// `Serialize` unless placed in a `Serializable<T>` wrapper, and thus they can be used to specify body section type.
     /// Please see the method below)
-    /// 
+    ///
     /// ## Example
     ///
     /// The string literal `"hello"` will be implicitly converted to `AmqpValue<&str>("hello")`
@@ -250,23 +250,23 @@ impl Sender {
     ///
     /// - Use `AmqpValue`, `AmqpSequence` or `Data` wrapper types to specify the body section type
     /// with all other message field set to None
-    /// 
+    ///
     /// ## Example
-    /// 
+    ///
     /// ```rust
     /// use fe2o3_amqp::types::primitives::Binary;
     /// use fe2o3_amqp::types::messaging::{AmqpValue, AmqpSequence, Data, Body};
-    /// 
+    ///
     /// // Send with Body::Value
     /// let outcome = sender.send(AmqpValue("hello world")).await.unwrap();
-    /// 
+    ///
     /// // Send with Body::Sequence
     /// let outcome = sender.send(AmqpSequence(vec![1i32, 2, 3])).await.unwrap();
-    /// 
+    ///
     /// // Send with Body::Data
     /// let outcome = sender.send(Data(Binary::from("hello world"))).await.unwrap();
     /// ```
-    /// 
+    ///
     /// - Use [`Body`] to specify the exact type of body section to use with all other message sections
     /// set to `None`
     ///