+++ serde_amqp/src/constants.rs
@@ -11,9 +11,12 @@ pub const DESCRIPTOR: &str = "AMQP1.0_DESCRIPTOR";
 #[doc(hidden)]
 pub const UNTAGGED_ENUM: &str = "FE2O3_AMQP_UNTAGGED";
 
+/// Use [`VALUE`] as the name if an enum needs to peek the format code before performing
+/// deserialization
+pub const VALUE: &str = "AMQP1.0_VALUE";
+
 // These constants should only be public on the crate level
 // to avoid accidental misuse
-pub(crate) const VALUE: &str = "AMQP1.0_VALUE";
 pub(crate) const ARRAY: &str = "AMQP1.0_ARRAY";
 pub(crate) const DECIMAL32: &str = "AMQP1.0_DECIMAL32";
 pub(crate) const DECIMAL64: &str = "AMQP1.0_DECIMAL64";