+++ serde_amqp/src/primitives/mod.rs
@@ -1,14 +1,18 @@
 //! Definition of the primitive types
 
 mod array;
+mod binary_ref;
 mod decimal;
+mod map;
 mod symbol;
 mod timestamp;
 mod uuid;
 
 // to avoid ambiguity
 pub use crate::primitives::array::*;
+pub use crate::primitives::binary_ref::*;
 pub use crate::primitives::decimal::*;
+pub use crate::primitives::map::*;
 pub use crate::primitives::symbol::*;
 pub use crate::primitives::timestamp::*;
 pub use crate::primitives::uuid::*;